\name{rforensicbatwing-package}
\alias{rforensicbatwing-package}
\alias{rforensicbatwing}
\docType{package}
\title{BATWING for calculating forensic trace-suspect match probabilities}
\description{A modified version (with great help from Ian J. Wilson) of Ian J. Wilson's program BATWING for calculating forensic trace-suspect match probabilities.}
\details{
Please see the documentation of \code{\link{coalmatchprob}} for an usage example.
}
\author{
Mikkel Meyer Andersen and Ian J. Wilson

Maintainer: Mikkel Meyer Andersen <mikl@math.aau.dk>
}
\references{
\href{http://www.mas.ncl.ac.uk/~nijw/batwing/batguide.pdf}{The BATWING manual}

I.J. Wilson, D.J. Balding, 
\emph{Genealogical inference from microsatellite data}, 
Genetics 150 (1998) 499-510.

I.J. Wilson, M.E. Weale, D.J. Balding, 
\emph{Inferences from DNA data: population histories, evolutionary processes and forensic match probabilities}, 
J. R. Stat. Soc. Ser. A 166 (2003) 155-201.

M.M. Andersen, A. Caliebe, A. Jochens, S. Willuweit, M. Krawczak, 
\emph{Estimating trace-suspect match probabilities for singleton Y-STR haplotypes using coalescent theory}, 
Forensic Sci. Int. Genet. (In Press, Corrected Proof \href{http://dx.doi.org/10.1016/j.fsigen.2012.11.004}{10.1016/j.fsigen.2012.11.004.}).
}
\keyword{ package }
\seealso{
\code{\link{batwing}}
\code{\link{coalmatchprob}}
}

